pyproject.toml
[tool.poetry]
name = "hooks"
version = "0.1.0"
description = ""
authors = ["AugustoFKL"]
readme = "README.md"

[tool.poetry.dependencies]
jira = "==3.5.2"
python-decouple = "==3.8"
python = "^3.11"
PyGithub = "==2.1.1"

[tool.poetry.group.dev.dependencies]
black = "==23.12.0"
commitizen = "^3.13.0"
pre-commit = "^3.6.0"
Sphinx = "==7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
